requirements-gpu.txt
tensorflow-gpu
numpy
opencv-python
lxml
tqdm
flask
seaborn
pillow
